requirements.txt
anyio==3.4.0
apischema==0.16.6
attrs==21.4.0
base58==2.1.1
bidict==0.21.4
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
construct==2.10.67
construct-typing==0.5.2
Flask==2.0.2
Flask-SocketIO==5.1.1
h11==0.12.0
httpcore==0.13.7
httpx==0.18.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
jsonrpcclient==4.0.2
jsonrpcserver==5.0.5
jsonschema==3.2.0
MarkupSafe==2.0.1
OSlash==0.6.3
pycparser==2.21
PyNaCl==1.4.0
pyrsistent==0.18.0
python-engineio==4.3.0
python-socketio==5.5.0
pyzmq==22.3.0
requests==2.26.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
solana==0.20.0
theblockchainapi==0.0.492
types-cachetools==4.2.7
typing-extensions==3.10.0.2
urllib3==1.26.7
websockets==10.1
Werkzeug==2.0.2
zmq==0.0.0
